Move instrumenter functions to separate file
